growth/dtc/requirements.txt
# Core
pydantic>=2.5.0
python-dotenv>=1.0.0

# HTTP & Async
httpx>=0.25.0
playwright>=1.40.0

# Data & Config
pyyaml>=6.0.1
jinja2>=3.1.0

# Logging
loguru>=0.7.0

# Dashboard
streamlit>=1.29.0

# Rate limiting helpers
tenacity>=8.2.0
